+++ test_terra_geocrud/settings/async.py
@@ -1,0 +1,3 @@
+from . import *  # NOQA
+
+GEOSTORE_EXPORT_CELERY_ASYNC = True